requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
langchain-qdrant>=0.1.0
pdfplumber>=0.10.0
qdrant-client>=1.7.0
pandas>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
tqdm>=4.66.0
openai>=1.0.0
sentence-transformers>=2.2.0

